day-trading-algo/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
yfinance>=0.1.70
joblib>=1.1.0
pyyaml>=6.0
requests>=2.27.0
tqdm>=4.62.0
pytz>=2021.3
python-dateutil>=2.8.2
